+++ lib/decimal.ex
@@ -2319,6 +2319,8 @@ defmodule Decimal do
 
   defp do_normalize(coef, exp), do: do_normalize_one(coef, exp)
 
+  defp do_normalize_one(0, _exp), do: %Decimal{coef: 0, exp: 0}
+
   defp do_normalize_one(coef, exp) when Kernel.rem(coef, 10) == 0 do
     do_normalize_one(Kernel.div(coef, 10), exp + 1)
   end
@@ -2336,6 +2338,8 @@ defmodule Decimal do
   end
 
   defp strip_trailing_zeros(coef, exp), do: strip_trailing_zeros_one(coef, exp)
+
+  defp strip_trailing_zeros_one(0, _exp), do: {0, 0}
 
   defp strip_trailing_zeros_one(coef, exp) when Kernel.rem(coef, 10) == 0 do
     strip_trailing_zeros_one(Kernel.div(coef, 10), exp + 1)